Revert "Merge tag 'android-6.0.1_r22' of https://android.googlesource.com/platform/frameworks/support into cm-13.0"
This reverts commit 98e06d73262bc654d546ef2ded2cd6c4db5b92bf, reversing
changes made to 6f0a2cb4415e2e170528c038a8f1cc2f68f60c3c.

TICKET: CYNGNOS-2250

+++ v7/gridlayout/src/android/support/v7/widget/GridLayout.java
@@ -1647,8 +1647,7 @@ public class GridLayout extends ViewGroup {
             boolean validSolution = true;
             // do a binary search to find the max delta that won't conflict with constraints
             while(deltaMin < deltaMax) {
-                // cast to long to prevent overflow.
-                final int delta = (int)(((long)deltaMin + deltaMax) / 2);
+                final int delta = (deltaMin + deltaMax) / 2;
                 invalidateValues();
                 shareOutDelta(delta, totalWeight);
                 validSolution = solve(getArcs(), a, false);
